requirement.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
defusedxml==0.7.1
Django==5.2.3
djangorestframework==3.16.0
httpie==3.2.4
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.4.4
Pygments==2.19.1
PySocks==1.7.1
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
setuptools==80.9.0
sqlparse==0.5.3
urllib3==2.4.0
